Fixes scripts for preparing release
